Cleanup login.ts file, rename to portal-account.ts
